Fix active schedule weekday matching

+++ services/commandService.js
@@ -40,7 +40,7 @@ async function isScheduleActiveNow(deviceId) {
   const dayName =
     new Intl.DateTimeFormat("en-US", {
       timeZone: "Asia/Kolkata",
-      weekday: "long"
+      weekday: "short"
     }).format(now);
 
   const result =
